backend/requirements.txt
# TaskFlow Backend Dependencies
# Python 3.11+

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database (SQLAlchemy 2.0)
sqlalchemy[asyncio]>=2.0.25
aiosqlite>=0.19.0

# PostgreSQL 支持 (预留迁移)
asyncpg>=0.29.0

# Validation & Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-dotenv>=1.0.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
email-validator>=2.1.0

# Scheduler
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.0
isort>=5.13.0
mypy>=1.8.0
